Deprecates Google AI JavaScript SDK
Deprecates the Google AI JavaScript SDK in favor of the new Google Gen AI SDK.

This change updates the README to reflect the deprecation and encourages users to migrate to the new SDK for the latest features and support.

All support for this repository, including bug fixes, will permanently end on August 31st, 2025.